platform_winit: move window loop into crate; root re-exports
- Replace facade with real winit loop using render_wgpu::gfx::Renderer
- Add deps to platform_winit and remove src/platform_winit.rs
- Root lib re-exports platform_winit crate
- Build/clippy/tests all green
