+++ Cargo.toml
@@ -31,19 +31,16 @@ gltf = { version = "1.4.1", features = ["import", "extensions"] }
 image = "0.25"
 log = "0.4.28"
 pollster = "0.4.0"
-rand = "0.9.2"
-rand_chacha = "0.9.0"
 serde = { version = "1.0.226", features = ["derive"] }
 serde_json = "1.0.145"
-serde_yaml = "0.9.34"
 wgpu = "26.0.1"
 which = "8.0.0"
 winit = "0.30.12"
 ra-assets = { version = "0.1.0", path = "shared/assets" }
 data_runtime = { version = "0.1.0", path = "crates/data_runtime" }
 sim_core = { version = "0.1.0", path = "crates/sim_core" }
-ux_hud = { version = "0.1.0", path = "crates/ux_hud" }
 render_wgpu = { version = "0.1.0", path = "crates/render_wgpu" }
+platform_winit = { version = "0.1.0", path = "crates/platform_winit" }
 
 [patch.crates-io]
 # Use a patched local copy of draco_decoder to avoid passing
